refactor tests to separate out the ones using deprecated logic. add test for new override behavior
